+++ iup/external/src/win/iupwin_tray.c
@@ -492,7 +492,7 @@ int iupdrvGetIconPixels(Ihandle* ih, const char* value, int* width, int* height,
   if (w <= 0 || h <= 0)
     return 0;
 
-  dst = (unsigned char*)malloc(w * h * 4);
+  dst = (unsigned char*)malloc((size_t)w * h * 4);
   if (!dst)
     return 0;
